backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Portfolio blog backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4"
]
